Split modules system, cli
Some parts of the CLI code is over-engineered, which is affecting new features. The `system` module is now the app domain (pure) while `cli` is the concrete user command to parse and convert into an operation. Separating `Operation` from `Command` flattens unused code when matching.
