+++ system/src/main.rs
@@ -4,7 +4,7 @@
 
 
 use clap::{Parser};
-use crate::main::system::System;
+use crate::main::cli::SystemCli;
 
 mod tmp;
 mod download;
@@ -13,13 +13,15 @@ mod cmd;
 mod image;
 mod package;
 mod os;
+
 mod main {
     pub mod system;
     pub mod exec;
+    pub mod cli;
 }
 
 fn main() {
-    let cli = System::parse();
+    let cli = SystemCli::parse();
     let exec = cli.operation.execute();
 
     match exec {
